Issuer notifications + security

+++ src/app/api/staff/annoucements/route.ts
@@ -8,77 +8,78 @@ import {
 } from "@/utils/response";
 import { Announcement } from "@prisma/client";
 import { createAnnouncementNotifications } from "@/utils/annoucement-notif";
+import { createNotification } from "@/utils/alerts";
 import { getServerToken } from "@/utils/session";
 
 export async function GET(
-    req: NextRequest,
-  ): Promise<NextResponse<ApiResponse<Announcement[] | null >>> {
-    try {
-      const token = await getServerToken(req);
-      if (!token || !token.id) return NextResponse.json(errorResponse(401, "Unauthorized"), { status: 401 });
-  
-      const id = token.id;
-  
-      console.log(id);
-  
-      if (!id) {
-        return NextResponse.json(errorResponse(400, "Id is required"), {
-          status: 400,
-        });
-      }
-  
-      const userProfile = await prisma.staff.findUnique({
-        where: {
-          id: String(id),
-        },
-      }) as unknown as { id: string; [key: string]: string | number | boolean | null };
-  
-      if (!userProfile) {
-        return NextResponse.json(errorResponse(404, "User profile not found"), {
-          status: 404,
-        });
-      }
-  
-      console.log("User Profile:", userProfile);
-  
-      const announcements = await prisma.announcement.findMany();
-  
-      const filteredAnnouncements = announcements.filter((announcement) => {
-        const filter = announcement.filter || {};
-        // if empty filter, return all announcements
-        if (Object.keys(filter).length === 0) {
-          return announcement.role === "STAFF";
-        }
-        // role checking
-        if (announcement.role !== "STAFF") {
-          return false;
-        }
-        return Object.entries(filter).every(
-          ([key, value]) => userProfile[key] === value
-        );
+  req: NextRequest,
+): Promise<NextResponse<ApiResponse<Announcement[] | null >>> {
+  try {
+    const token = await getServerToken(req);
+    if (!token || !token.id) return NextResponse.json(errorResponse(401, "Unauthorized"), { status: 401 });
+
+    const id = token.id;
+
+    console.log(id);
+
+    if (!id) {
+      return NextResponse.json(errorResponse(400, "Id is required"), {
+        status: 400,
       });
-  
-      return NextResponse.json(
-        successResponse(200, filteredAnnouncements, "Announcements fetched successfully"),
-        { status: 200 }
-      );
-    } catch (err) {
-      const error = err instanceof Error ? err.message : String(err);
-      return NextResponse.json(failureResponse(error), { status: 500 });
     }
+
+    const userProfile = await prisma.staff.findUnique({
+      where: {
+        id: String(id),
+      },
+    }) as unknown as { id: string; [key: string]: string | number | boolean | null };
+
+    if (!userProfile) {
+      return NextResponse.json(errorResponse(404, "User profile not found"), {
+        status: 404,
+      });
+    }
+
+    console.log("User Profile:", userProfile);
+
+    const announcements = await prisma.announcement.findMany();
+
+    const filteredAnnouncements = announcements.filter((announcement) => {
+      const filter = announcement.filter || {};
+      // if empty filter, return all announcements
+      if (Object.keys(filter).length === 0) {
+        return announcement.role === "STAFF" || announcement.issuer==id;
+      }
+      // role checking
+      if (announcement.role !== "STAFF") {
+        return false;
+      }
+      return Object.entries(filter).every(
+        ([key, value]) => userProfile[key] === value
+      );
+    });
+
+    return NextResponse.json(
+      successResponse(200, filteredAnnouncements, "Announcements fetched successfully"),
+      { status: 200 }
+    );
+  } catch (err) {
+    const error = err instanceof Error ? err.message : String(err);
+    return NextResponse.json(failureResponse(error), { status: 500 });
   }
+}
 
 export async function POST(
   req: NextRequest
 ): Promise<NextResponse<ApiResponse<Announcement | null>>> {
   try {
-    const { title, message, filter, role } = await req.json();
+    const token = await getServerToken(req);
+    if (!token || !token.id) return NextResponse.json(errorResponse(401, "Unauthorized"), { status: 401 });
 
-    if (role !== "STUDENT" && role !== "STAFF" && role !== "FACULTY") {
-      return NextResponse.json(errorResponse(400, "ACCESS DENIED, pls check roles"), {
-        status: 400,
-      });
-    }
+    const issuer = token.id;
+
+    const { title, message, filter, role } = await req.json();
+    const data = { title, message, filter, role, issuer: String(issuer) };
 
     if (!title || !message || !role) {
       return NextResponse.json(errorResponse(400, "Missing required fields"), {
@@ -86,16 +87,22 @@ export async function POST(
       });
     }
 
+    if (data.role!=="STUDENT" && data.role!=="FACULTY" && data.role!=="STAFF")
+    {
+      return NextResponse.json(errorResponse(400,"Missing permission"),{status:400});
+    }
+
     const announcement = await prisma.announcement.create({
-      data: { title, message, filter, role },
+      data,
     });
 
     const announcementWithCorrectFilter = {
-          ...announcement,
-          filter: announcement.filter as Record<string, string | number | boolean | Date | null>,
-        };
-        createAnnouncementNotifications(announcementWithCorrectFilter);
-        console.log("notification created");
+      ...announcement,
+      filter: announcement.filter as Record<string, string | number | boolean | Date | null>,
+    };
+    createAnnouncementNotifications(announcementWithCorrectFilter);
+    createNotification(String(token.id),"You created a new announcement",announcement.title);
+    console.log("notification created");
 
     return NextResponse.json(
       successResponse(201, announcement, "Announcement created successfully"),
@@ -107,28 +114,30 @@ export async function POST(
   }
 }
 
-
 export async function PATCH(req: NextRequest): Promise<NextResponse<ApiResponse<Announcement | null>>> {
   try {
+
+    const token = await getServerToken(req);
+    if (!token || !token.id) return NextResponse.json(errorResponse(401, "Unauthorized"), { status: 401 });
+
     const { id, ...data } = await req.json();
 
     if (!id) {
       return NextResponse.json(errorResponse(400, "Id is required"), { status: 400 });
     }
 
-    const findAnnouncement = await prisma.announcement.findUnique({
-      where: { id: String(id) },
-    });
+    const findAnnoucement = await prisma.announcement.findUnique({
+      where: {id : String(id)}
+    })
 
-    if (!findAnnouncement) {
-      return NextResponse.json(errorResponse(404, "Announcement not found"), {
-        status: 400,
-      });
+    if (!findAnnoucement)
+    {
+      return NextResponse.json(errorResponse(400,"Announcement not found"),{status: 400});
     }
-    if (findAnnouncement.role !== "STAFF" && findAnnouncement.role !== "STUDENT" && findAnnouncement.role !== "FACULTY") {
-        return NextResponse.json(errorResponse(400, "Cannot update this annoucement"), {
-            status: 400,
-        });
+
+    if (findAnnoucement.issuer!==token.id)
+    {
+      return NextResponse.json(errorResponse(400,"This annoucement was not issued by you"),{status:400});
     }
 
     const updatedAnnouncement = await prisma.announcement.update({
@@ -142,13 +151,29 @@ export async function PATCH(req: NextRequest): Promise<NextResponse<ApiResponse<
   }
 }
 
-
 export async function DELETE(req: NextRequest): Promise<NextResponse<ApiResponse<null>>> {
   try {
+    const token = await getServerToken(req);
+    if (!token || !token.id) return NextResponse.json(errorResponse(401, "Unauthorized"), { status: 401 });
+
     const { id } = await req.json();
 
     if (!id) {
       return NextResponse.json(errorResponse(400, "Id is required"), { status: 400 });
+    }
+
+    const findAnnoucement = await prisma.announcement.findUnique({
+      where: {id : String(id)}
+    })
+
+    if (!findAnnoucement)
+    {
+      return NextResponse.json(errorResponse(400,"Announcement not found"),{status: 400});
+    }
+
+    if (findAnnoucement.issuer!==token.id)
+    {
+      return NextResponse.json(errorResponse(400,"This annoucement was not issued by you"),{status:400});
     }
 
     await prisma.announcement.delete({
